pyproject.toml
[tool.poetry]
name = "plateaukit"
version = "0.1.0"
description = ""
authors = ["Kentaro Ozeki <32771324+ozekik@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "plateaukit" }]

[tool.poetry.dependencies]
python = "^3.8,<3.12"
cjio = { extras = ["reproject"], version = "^0.8.0" }
click = "^8.1.3"
tqdm = "^4.64.1"
geojson = "^2.5.0"
loguru = "^0.6.0"
pyproj = "^3.4.1"
lxml = "^4.9.2"
xmltodict = "^0.13.0"
tortoise-orm = "^0.19.2"
joblib = "^1.2.0"
aiomultiprocess = "^0.9.0"
geopandas = "^0.12.2"
rich = "^13.2.0"
prettytable = "^3.6.0"
requests = "^2.28.2"
xdg = "^5.1.1"
pydantic = "^1.10.4"
bidict = "^0.22.1"
pyogrio = "^0.5.1"
quantized-mesh-encoder = "^0.4.3"
rasterio = "^1.3.5.post1"
numpy = "<1.24"

[tool.poetry.scripts]
plateaukit = "plateaukit.cli:cli"

[tool.poetry.group.dev.dependencies]
pyglet = "^1.5.27"
scipy = "^1.10.0"
matplotlib = "^3.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
